Update test files.

+++ Test Files/parse_a_t5.c
@@ -1,4 +1,4 @@
-int square(int a)
+float square(float a)
 {
 	;
 }
